requirement.txt
langchain_groq
langchain_core
python-dotenv
streamlit
langchain_community
langchain_ollama
langserve
fastapi
uvicorn
langchain
sse_starlette
bs4
pypdf
chromadb
faiss-cpu